Adjusted the api
