+++ api_versions/v1.py
@@ -1,4 +1,4 @@
-from flask import Blueprint, request
+from flask import Blueprint, jsonify, request
 
 import detection
 from detection import DetectionData
@@ -11,6 +11,9 @@ logger = main_logger.getChild("v1")
 # Create a blueprint for this version of the API
 v1 = Blueprint("v1", import_name="v1")
 
+# The storage interface for the sessions
+session_storage = detection.session_storage
+
 
 # DEPRECATED
 @v1.route("/url", methods=["POST"])
@@ -20,3 +23,17 @@ def check_url_old():
     res = detection.test_old(DetectionData.from_json(json))
 
     return res.to_json_str_old()
+
+
+@v1.route("/url/state", methods=["POST"])
+def get_url_state():
+    json = request.get_json()
+    url = json["URL"]
+    uuid = json["uuid"]
+
+    session = session_storage.get_session(uuid, url)
+    status = session.get_state()
+
+    result = [{"status": status.result, "state": status.state}]
+
+    return jsonify(result)
